CMake: fix the shared library symlinks bug. #1
8 in https://github.com/KangLin/RabbitRemoteControl/issues/20

SOVERSION

ABI version number of a shared library target.

For shared libraries VERSION and SOVERSION can be used to specify the build version and ABI version respectively. When building or installing appropriate symlinks are created if the platform supports symlinks and the linker supports so-names. If only one of both is specified the missing is assumed to have the same version number. SOVERSION is ignored if NO_SONAME property is set.

A common convention is to specify both VERSION and SOVERSION such that SOVERSION matches the first component of VERSION:

See: https://cmake.org/cmake/help/latest/prop_tgt/SOVERSION.html#soversion